python_server/requirements.txt
# Python MCP Server Dependencies
# Built on FastMCP for rapid Tool integration

fastmcp>=0.1.0 # Or the relevant mcp package version currently supported
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
